add start scene

+++ Lava Vanguard/Assets/Scripts/Async/Bullet/BulletManager.cs
@@ -18,9 +18,10 @@ namespace Async
         {
             Vector3 spawnPos = PlayerManager.Instance.playerView.transform.position;
             int index = cardRankData.CardID[^1] - '1';
+            //Debug.Log("GenerateBullet - "+index+" mult: "+damageMultiplier);
             //Debug.Log(index);
             var b = Instantiate(bulletPrefabs[index], spawnPos, Quaternion.identity, bulletContainer).GetComponent<BulletView>();
-            b.damageMultiplier = (int)damageMultiplier;
+            b.damageMultiplier = damageMultiplier;
         }
     }
 }
